excide TabBar
вырезал TabBar
обновил вызов mainFlow в AppCoordinator
добавил функцию перезагрузки в CoordinatorFinishDelegate для обновления ListView при смене языка
